Add files via upload
A python script to search for relevant jobs on HackerNews. The HackerNews site is rich in terms of job postings but a little too basic for my taste(anybody's taste in fact) and do not have filters and search tools. Hence this script; to search for jobs based on keywords, over multiple pages of job posting.